### 2023-09-25 #5870 @editorialbot
[PRE REVIEW]: A new standard for drug repurposing by collaborative filtering: stanscofi and benchscofi
**Submitting author:** @clreda (<a href="http://orcid.org/0000-0003-3238-0258">Clémence Réda</a>)
**Repository:** https://github.com/RECeSS-EU-Project/stanscofi/
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.0
