+++ paper/paper.bib
@@ -83,6 +83,7 @@
   booktitle={2010 IEEE International conference on data mining},
   pages={995--1000},
   year={2010},
+  doi={10.1109/icdm.2010.127},
   organization={IEEE}
 }
 
@@ -217,7 +218,7 @@
     url = {\url{https://zenodo.org/record/7982969}},   
     author = {R\'{e}da, Cl\'{e}mence}, 
     year = {2023},   
-    doi = {110.5281/zenodo.7982969},
+    doi = {10.5281/zenodo.7982969},
     note = {Accessed on September 5th, 2023} 
 }
 
@@ -299,6 +300,7 @@
   number={8},
   pages={2538--2546},
   year={2020},
+  doi={10.1093/bioinformatics/btz965},
   publisher={Oxford University Press}
 }
 
