### 2023-09-26 #5870 @Kevin-Mattheus-Moerman
@clreda thanks for this submission. Can you clarify what this JOSS submission involves in therms of code reviewed.

Am I correct in saying that it involves the joint software contained in the following 2 repositories:
https://github.com/RECeSS-EU-Project/stanscofi
https://github.com/RECeSS-EU-Project/benchscofi

The former is listed here as the core repository for the paper etc and the `stanscofi` aspects. The software report :point_up: above reflects the content for this repository.

The latter is for `benchscofi` and running `cloc` locally provided.

```
Language                     files          blank        comment           code
Python                          35            647            800           3026
Jupyter Notebook                 4              0           3405            794
Markdown                         2            114              0            263
JSON                             2              0              0            113
YAML                             2              6              9             57
TOML                             1              0              0              3
SUM:                            46            767           4214           4256

```
