### 2023-09-26 #5870 @RECeSS-EU-Project
@Kevin-Mattheus-Moerman Thanks for managing the reviewing process!

Yes, the submission involves both _stanscofi_ and _benchscofi_ repositories.  The joint code allows to benchmark several drug repurposing approaches on several datasets, whereas automating as much as possible the model training and validation procedures. The reason why the code is split into two repositories is to increase modularity and usability of the code, as _stanscofi_ and _benchscofi_ serve different purposes. _stanscofi_ is targeted at the automation of the drug repurposing pipeline for model training and testing, whereas _benchscofi_ only aims at providing implementations of the state-of-the-art, and relies on the _stanscofi_ module.

Please let me know if the paragraph above answers your concerns.
