### 2023-10-18 #5870 @Nikoleta-v3
👋🏻 @singhkakan @darinddv @struckma @jaybee84 would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "A new standard for drug repurposing by collaborative filtering: stanscofi and benchscofi". You can find more information at the top of this Github issue 🆙
