### 2023-10-23 #5973 @editorialbot
[REVIEW]: stanscofi and benchscofi: a new standard for drug repurposing by collaborative filtering
**Submitting author:** @clreda (<a href="http://orcid.org/0000-0003-3238-0258">Clémence Réda</a>)
**Repository:** https://github.com/RECeSS-EU-Project/stanscofi/
**Branch with paper.md** (empty if default branch):
**Version:** v.2.0.0, v.2.0.0
