### 2023-11-11 #5973 @abhishektiwari
@clreda Initial comment on structure of Paper itself.

— Just for consistency purposes, the `Summary` should come before `Statement of need`.
—  When I read current `Summary` it does not describe the high-level functionality and purpose of the software for a diverse, non-specialist audience. See example paper [here](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
