### 2023-11-11 RECeSS-EU-Project/stanscofi#1 @abhishektiwari
Installation failed with conda
Installation failed. Using Mac M2 with Rossett mode on. Following [usage guide here](https://recess-eu-project.github.io/stanscofi/content.html#usage). Python `3.8.17`

Error log with `conda`

```
python3 -m pip install stanscofi
Collecting stanscofi
  Using cached stanscofi-2.0.0-py3-none-any.whl.metadata (7.0 kB)
Collecting pandas>=1.1.4 (from stanscofi)
  Using cached pandas-2.0.3-cp38-cp38-macosx_11_0_arm64.whl.metadata (18 kB)
Collecting numpy>=1.19.4 (from stanscofi)
  Using cached numpy-1.24.4-cp38-cp38-macosx_11_0_arm64.whl.metadata (5.6 kB)
Collecting scikit-learn>=0.24.0 (from stanscofi)
  Using cached scikit_learn-1.3.2-cp38-cp38-macosx_12_0_arm64.whl.metadata (11 kB)
Collecting scipy>=1.5.4 (from stanscofi)
  Using cached scipy-1.10.1-cp38-cp38-macosx_12_0_arm64.whl (28.8 MB)
Collecting matplotlib>=3.3.2 (from stanscofi)
  Using cached matplotlib-3.7.3-cp38-cp38-macosx_11_0_arm64.whl.metadata (5.7 kB)
Collecting threadpoolctl>=3.1.0 (from stanscofi)
  Using cached threadpoolctl-3.2.0-py3-none-any.whl.metadata (10.0 kB)
Collecting joblib>=1.0.1 (from stanscofi)
  Using cached joblib-1.3.2-py3-none-any.whl.metadata (5.4 kB)
Collecting tqdm>=4.58.0 (from stanscofi)
  Using cached tqdm-4.66.1-py3-none-any.whl.metadata (57 kB)
Collecting codecarbon>=2.2.2 (from stanscofi)
  Using cached codecarbon-2.3.1-py3-none-any.whl.metadata (6.5 kB)
Collecting seaborn>=0.11.0 (from stanscofi)
  Using cached seaborn-0.13.0-py3-none-any.whl.metadata (5.3 kB)
Collecting cute-ranking>=0.0.3 (from stanscofi)
  Using cached cute_ranking-0.0.3-py3-none-any.whl (8.7 kB)
Collecting umap-learn>=0.5.3 (from stanscofi)
  Using cached umap_learn-0.5.4-py3-none-any.whl
Collecting fastcluster>=1.2.6 (from stanscofi)
  Using cached fastcluster-1.2.6-cp38-cp38-macosx_11_0_arm64.whl (37 kB)
Collecting arrow (from codecarbon>=2.2.2->stanscofi)
  Using cached arrow-1.3.0-py3-none-any.whl.metadata (7.5 kB)
Collecting pynvml (from codecarbon>=2.2.2->stanscofi)
  Using cached pynvml-11.5.0-py3-none-any.whl (53 kB)
Collecting requests (from codecarbon>=2.2.2->stanscofi)
  Using cached requests-2.31.0-py3-none-any.whl.metadata (4.6 kB)
Collecting psutil (from codecarbon>=2.2.2->stanscofi)
  Using cached psutil-5.9.6-cp38-abi3-macosx_11_0_arm64.whl.metadata (21 kB)
Collecting py-cpuinfo (from codecarbon>=2.2.2->stanscofi)
  Using cached py_cpuinfo-9.0.0-py3-none-any.whl (22 kB)
Collecting fuzzywuzzy (from codecarbon>=2.2.2->stanscofi)
  Using cached fuzzywuzzy-0.18.0-py2.py3-none-any.whl (18 kB)
Collecting click (from codecarbon>=2.2.2->stanscofi)
  Using cached click-8.1.7-py3-none-any.whl.metadata (3.0 kB)
Collecting prometheus-client (from codecarbon>=2.2.2->stanscofi)
  Using cached prometheus_client-0.18.0-py3-none-any.whl.metadata (26 kB)
Collecting numpy>=1.19.4 (from stanscofi)
  Using cached numpy-1.19.4.zip (7.3 MB)
  Installing build dependencies ... done
  Preparing metadata (pyproject.toml) ... error
  error: subprocess-exited-with-error

  × Preparing metadata (pyproject.toml) did not run successfully.
  │ exit code: 1
  ╰─> [180 lines of output]
      Processing numpy/random/_bounded_integers.pxd.in
      Processing numpy/random/_philox.pyx
      Processing numpy/random/_bounded_integers.pyx.in
      Processing numpy/random/_sfc64.pyx
      Processing numpy/random/_mt19937.pyx
      Processing numpy/random/bit_generator.pyx
      Processing numpy/random/mtrand.pyx
      Processing numpy/random/_generator.pyx
      Processing numpy/random/_pcg64.pyx
      Processing numpy/random/_common.pyx
      Cythonizing sources
      blas_opt_info:
      blas_mkl_info:
      customize UnixCCompiler
        libraries mkl_rt not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      blis_info:
        libraries blis not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      openblas_info:
        libraries openblas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      atlas_3_10_blas_threads_info:
      Setting PTATLAS=ATLAS
        libraries tatlas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      atlas_3_10_blas_info:
        libraries satlas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      atlas_blas_threads_info:
      Setting PTATLAS=ATLAS
        libraries ptf77blas,ptcblas,atlas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      atlas_blas_info:
        libraries f77blas,cblas,atlas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      accelerate_info:
        libraries accelerate not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
      Library accelerate was not found. Ignoring
        libraries veclib not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
      Library veclib was not found. Ignoring
        FOUND:
          extra_compile_args = ['-msse3', '-I/System/Library/Frameworks/vecLib.framework/Headers']
          extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
          define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

        FOUND:
          extra_compile_args = ['-msse3', '-I/System/Library/Frameworks/vecLib.framework/Headers']
          extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
          define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

      non-existing path in 'numpy/distutils': 'site.cfg'
      lapack_opt_info:
      lapack_mkl_info:
        libraries mkl_rt not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      openblas_lapack_info:
        libraries openblas not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      openblas_clapack_info:
        libraries openblas,lapack not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      flame_info:
        libraries flame not found in ['/opt/homebrew/anaconda3/envs/stanscofi_env/lib', '/usr/lib']
        NOT AVAILABLE

      atlas_3_10_threads_info:
      Setting PTATLAS=ATLAS
        libraries lapack_atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries tatlas,tatlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries lapack_atlas not found in /usr/lib
        libraries tatlas,tatlas not found in /usr/lib
      <class 'numpy.distutils.system_info.atlas_3_10_threads_info'>
        NOT AVAILABLE

      atlas_3_10_info:
        libraries lapack_atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries satlas,satlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries lapack_atlas not found in /usr/lib
        libraries satlas,satlas not found in /usr/lib
      <class 'numpy.distutils.system_info.atlas_3_10_info'>
        NOT AVAILABLE

      atlas_threads_info:
      Setting PTATLAS=ATLAS
        libraries lapack_atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries ptf77blas,ptcblas,atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries lapack_atlas not found in /usr/lib
        libraries ptf77blas,ptcblas,atlas not found in /usr/lib
      <class 'numpy.distutils.system_info.atlas_threads_info'>
        NOT AVAILABLE

      atlas_info:
        libraries lapack_atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries f77blas,cblas,atlas not found in /opt/homebrew/anaconda3/envs/stanscofi_env/lib
        libraries lapack_atlas not found in /usr/lib
        libraries f77blas,cblas,atlas not found in /usr/lib
      <class 'numpy.distutils.system_info.atlas_info'>
        NOT AVAILABLE

        FOUND:
          extra_compile_args = ['-msse3', '-I/System/Library/Frameworks/vecLib.framework/Headers']
          extra_link_args = ['-Wl,-framework', '-Wl,Accelerate']
          define_macros = [('NO_ATLAS_INFO', 3), ('HAVE_CBLAS', None)]

      running dist_info
      running build_src
      build_src
      building py_modules sources
      creating build
      creating build/src.macosx-11.1-arm64-3.8
      creating build/src.macosx-11.1-arm64-3.8/numpy
      creating build/src.macosx-11.1-arm64-3.8/numpy/distutils
      building library "npymath" sources
      ld: warning: ignoring file '/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/_configtest.o': found architecture 'arm64', required architecture 'x86_64'
      ld: Undefined symbols:
        _main, referenced from:
            <initial-undefines>
      clang: error: linker command failed with exit code 1 (use -v to see invocation)
      Running from numpy source directory.
      setup.py:480: UserWarning: Unrecognized setuptools command, proceeding with generating Cython sources and expanding templates
        run_build = parse_setuppy_commands()
      /private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/dist.py:275: UserWarning: Unknown distribution option: 'define_macros'
        warnings.warn(msg)
      Traceback (most recent call last):
        File "/opt/homebrew/anaconda3/envs/stanscofi_env/lib/python3.8/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 353, in <module>
          main()
        File "/opt/homebrew/anaconda3/envs/stanscofi_env/lib/python3.8/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "/opt/homebrew/anaconda3/envs/stanscofi_env/lib/python3.8/site-packages/pip/_vendor/pyproject_hooks/_in_process/_in_process.py", line 149, in prepare_metadata_for_build_wheel
          return hook(metadata_directory, config_settings)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/build_meta.py", line 157, in prepare_metadata_for_build_wheel
          self.run_setup()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/build_meta.py", line 248, in run_setup
          super(_BuildMetaLegacyBackend,
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/build_meta.py", line 142, in run_setup
          exec(compile(code, __file__, 'exec'), locals())
        File "setup.py", line 508, in <module>
          setup_package()
        File "setup.py", line 500, in setup_package
          setup(**metadata)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/core.py", line 169, in setup
          return old_setup(**new_attr)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/__init__.py", line 165, in setup
          return distutils.core.setup(**attrs)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/core.py", line 148, in setup
          dist.run_commands()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/dist.py", line 967, in run_commands
          self.run_command(cmd)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/dist.py", line 986, in run_command
          cmd_obj.run()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/command/dist_info.py", line 31, in run
          egg_info.run()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/command/egg_info.py", line 24, in run
          self.run_command("build_src")
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/cmd.py", line 313, in run_command
          self.distribution.run_command(command)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-build-env-ely895lm/overlay/lib/python3.8/site-packages/setuptools/_distutils/dist.py", line 986, in run_command
          cmd_obj.run()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/command/build_src.py", line 144, in run
          self.build_sources()
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/command/build_src.py", line 155, in build_sources
          self.build_library_sources(*libname_info)
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/command/build_src.py", line 288, in build_library_sources
          sources = self.generate_sources(sources, (lib_name, build_info))
        File "/private/var/folders/jc/7dkljn7n3hb2w2jsw27d0ghh0000gn/T/pip-install-x1dc175v/numpy_dcf5fa9286b1418c9d7fded1b62dc772/numpy/distutils/command/build_src.py", line 378, in generate_sources
          source = func(extension, build_dir)
        File "numpy/core/setup.py", line 663, in get_mathlib_info
          raise RuntimeError("Broken toolchain: cannot link a simple C program")
      RuntimeError: Broken toolchain: cannot link a simple C program
      [end of output]

error: metadata-generation-failed

× Encountered error while generating package metadata.
╰─> See above for output.

note: This is an issue with the package mentioned above, not pip.
hint: See above for details.
```
