### 2023-11-11 #5973 @abhishektiwari
@clreda I spent a few hours trying to install the package but neither `pip` nor `conda` instructions work (details added on the GitHub issue [here](https://github.com/RECeSS-EU-Project/stanscofi/issues/1)). I use Mac M2 laptop and tried running these with `Rosetta` mode but no luck. Is there a `Dockerfile` one can use to test these packages?

Can you share what OS is supported by stanscofi?
