### 2023-11-11 RECeSS-EU-Project/stanscofi#2 @abhishektiwari
Instructions to run stanscofi as a Docker container
Instructions to anyone wanting run `stanscofi` as docker container.

1. Add `Dockerfile` root of the `stanscofi` repository

```
# Pull base image.
FROM python:3.8.18-bullseye

# Install.
RUN \
  apt-get update && \
  apt-get -y upgrade && \
  apt-get install -y build-essential && \
  apt-get install -y software-properties-common && \
  apt-get install -y byobu curl git htop man unzip vim wget pip && \
  pip install stanscofi notebook

# Add files.
ADD . /home/stanscofi

# Set environment variables.
ENV HOME /home

# Define working directory.
WORKDIR /home/stanscofi

# Define default command.
CMD ["bash"]

EXPOSE 3000
```

2. Build Docker image in root of the `stanscofi` repository

```
docker build -t stanscofi .
```

3. Run Docker image built in previous step and drop into SSH
```
docker run -it --expose 3000  -p 3000:3000 stanscofi
```

4. Run notebook
```
cd docs
jupyter notebook --ip 0.0.0.0 --no-browser --allow-root --port 3000
```

5. Access notebook at `http://127.0.0.1:3000/tree`
