### 2023-11-15 ff31299
<paper> shorten, rewrite structure
