+++ paper/paper.md
@@ -29,25 +29,23 @@ date: 5 September 2023
 bibliography: paper.bib
 ---
 
+# Summary
+
+As of 2023, current drug development pipelines last around ten years, costing $2.3 billion on average [@philippidis2023unbearable], while drug commercialization failure rates go up to 90% [@sun202290]. Drug repurposing might mitigate these issues by speeding up the drug discovery phase on well-documented compounds [@jarada2020review], helping to prevent adverse side effects and low accrual in clinical trials [@hingorani2019improving]. Recent papers [@zhang2017computational;@yang2019additional;@he2020hybrid;@meng2022weighted;@yang2022computational;@yang2023self] have reported near-perfect predicting power (*area under the curve*, or AUC) on several repurposing datasets by resorting to collaborative filtering approaches. Collaborative filtering straightforwardly allows the implementation of sparse classifiers which aggregate the information from many diseases. However, a considerable hurdle to developing efficient drug repurposing approaches based on collaborative filtering is the lack of a standard pipeline to train, validate and compare these algorithms on a robust dataset.
+
 # Statement of need
-
-As of 2023, current drug development pipelines last around ten years, costing $2.3 billion on average [@philippidis2023unbearable], while drug commercialization failure rates go up to 90% [@sun202290]. Drug repurposing might mitigate these issues. Drug repurposing methods [@jarada2020review] systematically screen chemical compounds for new therapeutic indications. Indeed, documentation about those compounds is available and helps prevent adverse side effects and low accrual in clinical trials. These are still the main reasons for failure in late clinical phases [@hingorani2019improving]. Some prior works implement drug repurposing through collaborative filtering [@zhang2017computational;@meng2022weighted;@yang2019additional]. This semi-supervised learning framework leverages known drug-disease matchings to recommend new ones. Predicted drug-disease associations stem from a function whose parameters are learnt on a whole matrix of drug-disease matchings instead of focusing on a single disease at a time. In particular, literature often relies on tensor decomposition, *i*.*e*., any drug-disease matching in the matrix is the output of a classifier in which only lower-rank tensors intervene, *e*.*g*., factorization machines (FMs) [@rendle2010factorization]. Then, collaborative filtering straightforwardly allows the implementation of sparse models which aggregate the information from many diseases. That might account for the popularity of collaborative filtering approaches to drug repurposing.
-
-Indeed, recent works [@he2020hybrid;@yang2022computational;@yang2023self] have reported near-perfect predicting power (*area under the curve*, or AUC) on several repurposing datasets. However, a considerable hurdle to developing efficient drug repurposing approaches based on collaborative filtering is the lack of a standard pipeline to train, validate and compare these algorithms on a robust dataset.
 
 The **stanscofi** Python package [@reda2023stanscofi] comprises method-agnostic training and validation procedures on several public drug repurposing datasets. Implementing these steps is crucial to avoid data leakage, *i*.*e*. the model is learnt over information that should be unavailable at prediction time. Indeed, data leakage is the source of a significant reproducibility crisis in machine learning [@roelofs2019meta;@feldman2019advantages;@kapoor2023leakage]. Our package avoids data leakage in two ways: first, by building weakly correlated training and validation sets for the drug feature vectors, and second, by implementing a generic model class, which allows the automation of the training and validation procedures.
 
[…]
