+++ .github/workflows/gh-pages.yml
@@ -1,8 +1,9 @@
 name: Release documentation
 
-on:
-  release:
-    types: [published]
+on: push
+#on:
+#  release:
+#    types: [published]
 
 jobs:
   build:
