### 2023-11-15 RECeSS-EU-Project/stanscofi#1 @RECeSS-EU-Project
Hi @abhishektiwari, it looks like several libraries needed to compile C code (and install the numpy package, which is the point where the installation fails) are missing on your computer.

Please ensure that you can install numpy on your computer prior to installing stanscofi, for instance by running the following lines:
```bash
conda create -n test_env python=3.8.5 -y
conda activate test_env
python3 -m pip install numpy==1.19.4
```
and try to fix the compilation errors, for instance, by checking that you have installed Xcode and/or gcc files are available (and reachable) on your computer.
