### 2023-11-15 RECeSS-EU-Project/stanscofi#2 @RECeSS-EU-Project
Hi @abhishektiwari, this is a very nice addition to the guidelines to improve stanscofi usability. Thank you very much, I will add the Dockerfile and instructions.
