### 2023-11-18 #5973 @abhishektiwari
Thanks, @clreda I have now worked my way through the `stanscofi` and was able to validate software functionality.

I am now having issue with [benchscofi](https://github.com/RECeSS-EU-Project/benchscofi) due it's restrictive `tensorflow` requirements.

Created ticket here: https://github.com/RECeSS-EU-Project/benchscofi/issues/1
