### 2023-11-21 #5973 @clreda
I have replied to your issue. I believe the best course of action is to provide for users the alternative installation guide, which relies on pip dependency conflict solver, and keep the current one to ensure smooth automated tests. Please feel free to mention other possibilities which might improve user functionality
