### 2023-12-05 #5973 @Nikoleta-v3
Hello @abhishektiwari @jaybee84👋🏻, any updates regarding your reviews?

@clreda, I can see that https://github.com/RECeSS-EU-Project/stanscofi/issues/2 is now closed. In the future, could you indicate the commit hash (or even better, the pull request) that resolved the issue? It will make reviewers' life a bit easier 😄
