### 2023-12-06 #5973 @abhishektiwari
@clreda Looked at revised paper structure, still not sold on current summary. I generally think software paper summary as an elevator pitch - problem statement, solution, key value proposition, any additional detail for target audience. Revised version mentions the problem statement but not the solution.
