### 2023-12-06 RECeSS-EU-Project/stanscofi#3 @jaybee84
pip installation not working 
As a part of review (https://github.com/openjournals/joss-reviews/issues/5973) I tried installing with both `pip` and `pip3` and both were unsuccessful. Please see error logs below:

```
pip3 install stanscofi

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
cnvkit 0.9.9 requires joblib<1.0, but you have joblib 1.3.2 which is incompatible.
Successfully installed numpy-1.19.4
```

```
pip install stanscofi

ERROR: pip's dependency resolver does not currently take into account all the packages that are installed. This behaviour is the source of the following dependency conflicts.
cnvkit 0.9.9 requires joblib<1.0, but you have joblib 1.3.2 which is incompatible.
Successfully installed arrow-1.3.0 codecarbon-2.3.2 cute-ranking-0.0.3 fastcluster-1.2.6 fuzzywuzzy-0.18.0 joblib-1.3.2 llvmlite-0.39.1 numba-0.56.4 numpy-1.21.3 prometheus-client-0.19.0 py-cpuinfo-9.0.0 pynndescent-0.5.11 pynvml-11.5.0 scikit-learn-1.3.2 seaborn-0.12.2 stanscofi-2.0.0 threadpoolctl-3.2.0 types-python-dateutil-2.8.19.14 umap-learn-0.5.5
```

I am running this on `macOSMonterey` and a conda environment with `Python 3.8.12`
