### 2023-12-06 RECeSS-EU-Project/stanscofi#3 @jaybee84
looks like it is an issue with my system having a higher version of `joblib`. Given that many people might have this version, is it possible to update the version of `cnvkit` and `joblib` used? This may be helpful in increasing the longevity of the package as the dependencies update over time.
