+++ paper/paper.md
@@ -31,19 +31,19 @@ bibliography: paper.bib
 
 # Summary
 
-As of 2023, current drug development pipelines last around ten years, costing $2.3 billion on average [@philippidis2023unbearable], while drug commercialization failure rates go up to 90% [@sun202290]. Drug repurposing might mitigate these issues by speeding up the drug discovery phase on well-documented compounds [@jarada2020review], helping to prevent adverse side effects and low accrual in clinical trials [@hingorani2019improving]. Recent papers [@zhang2017computational;@yang2019additional;@he2020hybrid;@meng2022weighted;@yang2022computational;@yang2023self] have reported near-perfect predicting power (*area under the curve*, or AUC) on several repurposing datasets by resorting to collaborative filtering approaches. Collaborative filtering straightforwardly allows the implementation of sparse classifiers which aggregate the information from many diseases. However, a considerable hurdle to developing efficient drug repurposing approaches based on collaborative filtering is the lack of a standard pipeline to train, validate and compare these algorithms on a robust dataset.
+Drug development is still a time-consuming and costly process as of today, while the failure rate in the successful commercialization of drug candidates is high. Drug repurposing is an approach which screens currently available chemical compounds and tool molecules to uncover novel therapeutic indications. In particular, collaborative filtering has sparked interest, as this framework allows us to deal with implicit information on drug-disease associations. As popular as drug repurposing might be, the lack of standard training, validation pipelines and benchmark datasets hinders the development and assessment of drug repurposing methods. To overcome this issue, we propose Python package **stanscofi** (*STANdard for drug Screening in COllaborative FIltering*), which permits the quick implementation of ready-to-go drug repurposing models and ensures proper training and validation of the methods. We also built the Python package **benchscofi** (*BENCHmark for drug Screening in COllaborative FIltering*) upon **stanscofi** to implement several algorithms from the state-of-the-art and enable the first large-scale benchmark of the field.
 
 # Statement of need
 
-The **stanscofi** Python package [@reda2023stanscofi] comprises method-agnostic training and validation procedures on several public drug repurposing datasets. Implementing these steps is crucial to avoid data leakage, *i*.*e*. the model is learnt over information that should be unavailable at prediction time. Indeed, data leakage is the source of a significant reproducibility crisis in machine learning [@roelofs2019meta;@feldman2019advantages;@kapoor2023leakage]. Our package avoids data leakage in two ways: first, by building weakly correlated training and validation sets for the drug feature vectors, and second, by implementing a generic model class, which allows the automation of the training and validation procedures.
+As of 2023, current drug development pipelines last around ten years, costing $2.3 billion on average [@philippidis2023unbearable], while drug commercialization failure rates go up to 90% [@sun202290]. Drug repurposing might mitigate these issues by speeding up the drug discovery phase on well-documented compounds [@jarada2020review], helping to prevent adverse side effects and low accrual in clinical trials [@hingorani2019improving]. Recent papers [@zhang2017computational;@yang2019additional;@he2020hybrid;@meng2022weighted;@yang2022computational;@yang2023self] have reported near-perfect predicting power (*area under the curve*, or AUC) on several repurposing datasets by resorting to collaborative filtering approaches. Collaborative filtering straightforwardly allows the implementation of sparse classifiers which aggregate the information from many diseases. However, a considerable hurdle to developing efficient drug repurposing approaches based on collaborative filtering is the lack of a standard pipeline to train, validate and compare these algorithms on a robust dataset. 
 
[…]
