### 2023-12-06 #5973 @clreda
@abhishektiwari Thanks for your feedback! Please find enclosed the newest revision of the paper. Let me know if that tackles your concerns. (commit [196bd5f](https://github.com/RECeSS-EU-Project/stanscofi/commit/196bd5ff9c35408c9eea61a83a0abecb41788078))

@jaybee84 Thanks for accepting to review this submission! Package ``cnvkit`` is not part of the requirements of packages ``stanscofi`` nor ``benchscofi``.  The only constraints version-wise on these packages is that ``joblib>=1.0.1``, which allows the use of later versions of ``joblib``.
