### 2023-12-06 RECeSS-EU-Project/stanscofi#3 @RECeSS-EU-Project
Hi! Package ``cnvkit`` is not part of the requirements of packages ``stanscofi`` nor ``benchscofi``. The only constraints version-wise is that ``joblib>=1.0.1``, so I suspect the issue actually comes from ``cnvkit``. Either way, this issue flag from ``pip`` should not prevent you from running the package.

Let me know if you encounter an issue when using the package. Otherwise, I will close the issue.
