### 2023-12-09 #5973 @abhishektiwari
@clreda thanks for revising the summary.

**Additional feedback on installation:**

At this point my only non-blocking concern is that installation of `benchscofi` package is not easy. Due to a large number of dependencies it took a fair bit of my time to install `benchscofi`. This may discourage people trying to use these packages.

I strongly suggest providing a Docker images for both `stanscofi` and `benchscofi`.

Problem become more obvious when you see PyPI release pipeline for `benchscofi` never succeeded. Author should fix PyPI and add Conda releases for `benchscofi`. Release pipeline for `stanscofi` seems better but similar issues. Last build succeeded for PyPI and failed for Conda.

** Finally **
Overall, `stanscofi` and `benchscofi` are very useful packages to anyone working on drug screening and repurposing. `stanscofi` offers end to end workflow for screening whereas `benchscofi` implements a good number of published algorithms.

@Nikoleta-v3 Completed my review.
