### 2023-12-14 #5973 @jaybee84
@clreda I am having trouble reviewing the functionality of stanscofi.

After following the instructions to start a jupyter notebook, I tried to run the notebook cells and none of the dependencies were available. The kernel running the notebook was the preferred kernel of python3 started within the conda environment.

Can you please improve the instructions [here](https://github.com/RECeSS-EU-Project/stanscofi/) to walk a new user through all steps of understanding the functionality and setup for using this package?
