### 2023-12-14 #5973 @jaybee84
@clreda I am trying to review the functionality of benchscofi and am having some trouble. I strongly recommend adding documentation about benchscofi that is similar to the jupyter notebook for stanscofi. Once that is added, I should be able to complete my review.
