### 2023-12-14 #5973 @clreda
Dear @abhishektiwari, I agree with your point on the installation of benchscofi. There are several cross-platform algorithms, hence the large number of dependencies. I will work on Docker images for both packages, and ensure that the pipelines for PyPI and Conda work for stanscofi and benchscofi in the next weeks. The main issue with including a conda release for benchscofi is that some of the dependencies are not present in conda and would require a script file (ex. build.sh) to download them from PyPI and install them locally during the installation (which I will try to work on in the next weeks).

Dear @jaybee84 Thanks for your feedback! I assume that you managed to test the functionality of stanscofi. May I close the following issue ([issue #3](https://github.com/RECeSS-EU-Project/stanscofi/issues/3))? I don't understand whether the issue with benchscofi is about the installation (and having a similar list of commands to execute benchscofi within Docker and run the notebooks) or having access to an introductory notebook in the first place (in that case, there are several notebooks in the [docs/](https://github.com/RECeSS-EU-Project/benchscofi/tree/master/docs) folder: in particular, one for using algorithms within benchscofi ([here](https://github.com/RECeSS-EU-Project/benchscofi/blob/master/docs/RankingMetrics.ipynb)) and another one on class prior estimations ([here](https://github.com/RECeSS-EU-Project/benchscofi/blob/master/docs/Class%20prior%20estimation.ipynb))). Please let me know what is the problem.
