### 2023-12-15 RECeSS-EU-Project/stanscofi#3 @jaybee84
I was able to install the package.
