### 2023-12-15 #5973 @jaybee84
Ah! @clreda I see that the issue is that I was not looking at the right place. Since both packages are being released in the same publication, I expected that there would be a pointer from the stanscofi repo to the benchscofi repo so that all documentation for both packages are easy to find from one place. I see now that the `RECeSS EU project` is the common origin point to navigate to both packages. I will take a look at the benchscofi functionality soon.
