### 2023-12-20 #5973 @RECeSS-EU-Project
Dear @jaybee84 thanks for your contribution. I have merged the PR. (commit [0f960a26](https://github.com/RECeSS-EU-Project/benchscofi/commit/0f960a26e8e4ba33911907548e74b0ce97bb2d43))
