+++ .github/workflows/conda-release.yml
@@ -1,29 +1,10 @@
-name: stancofi release
+name: stancofi conda release
 
 on:
   release:
     types: [published]
 
 jobs:
-  pypi-publish:
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v2
-    - run: sed -i "s:9999:${VERSION//*v/}:" setup.py conda/meta.yaml
-      env:
-          VERSION: ${{ github.ref }}
-    - uses: actions/setup-python@v1
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install setuptools wheel twine
-    - name: Build and publish
-      env:
-        TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
-        TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
-      run: |
-        python setup.py sdist bdist_wheel
-        twine upload dist/*
   conda-publish:
     runs-on: ubuntu-latest
     steps:
