### 2024-01-04 c94bf1a
<docs> add instructions for docker image
