### 2024-01-04 #5973 @clreda
Dear @abhishektiwari Docker images for [stanscofi](https://hub.docker.com/repository/docker/recessproject/stanscofi) and [benchscofi](https://hub.docker.com/repository/docker/recessproject/benchscofi) are now available on DockerHub. The documentation in both repositories has been updated accordingly. The release pipelines have been fixed (stanscofi: commit [981bef4](https://github.com/RECeSS-EU-Project/stanscofi/commit/981bef42a41e3fe084d34e397edf50fad142ed90); benchscofi: commit [b41d2d9](https://github.com/RECeSS-EU-Project/benchscofi/commit/b41d2d966e01486903c273ded76947254935fdb0)). Only thing that is missing is the Conda release pipeline for benchscofi.
