### 2024-01-07 #5973 @Nikoleta-v3
Thank you for your patience @clreda, and thanks to the reviewers for their time! I believe both their comments greatly benefited the paper, especially regarding installation.

**Documentation:**

Regarding documentation, I still think something is missing. I realize that the packages have two separate repositories for a purpose, but have you ever considered having a common Read the Docs?

My issue is that https://recess-eu-project.github.io/stanscofi/ has very minimal information. Everything in [Introduction to stanscofi.ipynb](https://github.com/RECeSS-EU-Project/stanscofi/blob/master/docs/Introduction%20to%20stanscofi.ipynb) should be part of the online documentation. I shouldn't have to clone the repo to see the examples. You can open the Jupyter notebook on GitHub, but then I can't navigate to the sections I want by clicking. So, it just seems a bit too much for the user end. The `benchscofi` package doesn't have online documentation but has a few Jupyter notebooks, which is why I am suggesting a common online document. What do you think?

FYI, it doesn't have to be on Read the Docs. I see that you also have this: https://recess-eu-project.github.io/. It could be anywhere. At the moment, the [online documentation](https://recess-eu-project.github.io/stanscofi)  just doesn't add much.

**Other comments:**

- In `benchscofi`, you have a contributions section 👍🏻 Are you expecting contributions to implement tests for their contributions? If so, can you add some information. Namely, that you use `unittest` and details on how to run the test suite locally.

- Regarding the paper, there is a small typo in line 16. "we propose **the**"
