+++ docs/models.rst
@@ -1,9 +1,6 @@
 Models
 =================
 
-stanscofi.models module
------------------------
-
 .. automodule:: stanscofi.models
    :members:
    :undoc-members:
