+++ docs/trainingtesting.rst
@@ -1,9 +1,6 @@
 Training/Testing
 =================
 
-stanscofi.training\_testing module
-----------------------------------
-
 .. automodule:: stanscofi.training_testing
    :members:
    :undoc-members:
