+++ docs/utils.rst
@@ -1,9 +1,6 @@
 Utils
 =================
 
-stanscofi.utils module
-----------------------
-
 .. automodule:: stanscofi.utils
    :members:
    :undoc-members:
