+++ README.md
@@ -10,84 +10,25 @@ This repository is a part of the EU-funded [RECeSS project](https://recess-eu-pr
 
 As of 2022, current drug development pipelines last around 10 years, costing $2billion in average, while drug commercialization failure rates go up to 90%. These issues can be mitigated by drug repurposing, where chemical compounds are screened for new therapeutic indications in a systematic fashion. In prior works, this approach has been implemented through collaborative filtering. This semi-supervised learning framework leverages known drug-disease matchings in order to recommend new ones.
 
-The **stanscofi** package comprises method-agnostic training, validation, preprocessing and visualization procedures on several published drug repurposing datasets. The proper implementation of these steps is crucial in order to avoid data leakage, *i*.*e*., the model is learnt over information that should be unavailable at prediction time. Indeed, data leakage is the source of a major reproducibility crisis in machine learning. This will be avoided by building training and validation sets which are weakly correlated with respect to the drug and disease feature vectors. The main performance metric will be the area under the curve (AUC), which estimates the diagnostic ability of a recommender system better than accuracy in imbalanced datasets.
+The *stanscofi* package comprises method-agnostic training, validation, preprocessing and visualization procedures on several published drug repurposing datasets. The proper implementation of these steps is crucial in order to avoid data leakage, *i*.*e*., the model is learnt over information that should be unavailable at prediction time. Indeed, data leakage is the source of a major reproducibility crisis in machine learning. This will be avoided by building training and validation sets which are weakly correlated with respect to the drug and disease feature vectors. The main performance metric will be the area under the curve (AUC), which estimates the diagnostic ability of a recommender system better than accuracy in imbalanced datasets.
 
 Medium-term outcomes to this package will significantly alleviate the economic burden of drug discovery pipelines, and will help find treatments in a more sustainable manner, especially for rare or tropical neglected diseases.
 
-For more information about the datasets accessible in **stanscofi**, please refer to the following [repository](https://github.com/RECeSS-EU-Project/drug-repurposing-datasets).
+For more information about the datasets accessible in *stanscofi*, please refer to the following [repository](https://github.com/RECeSS-EU-Project/drug-repurposing-datasets).
 
 ## Install the latest release
 
-### Using pip (package hosted on PyPI)
+Run one of the following commands:
 
 ```bash
+## Using pip: install in Python env
 pip install stanscofi
-```
-
-### Using conda (package hosted on Anaconda.org)
-
-```bash
+## Using Anaconda: install in Python env
 conda install -c recess stanscofi
+## Using the Docker image: will open a container
+docker push recessproject/stanscofi:2.0.0
 ```
-
-### Using Docker image (image hosted on DockerHub)
-
-```bash
-docker push recessproject/stanscofi:<version>
-```
-
-## Running notebooks using Docker
-
-Credits to [Abhishek Tiwari](https://github.com/abhishektiwari) for the Dockerfile, instructions and comments. In the root folder of the repository, run the following commands
-
-```bash
-#Build Docker image
-docker build -t stanscofi .
-#Run Docker image built in previous step and drop into SSH
-docker run -it --expose 3000  -p 3000:3000 stanscofi 
-#Run notebook
-cd docs/ && jupyter notebook --ip 0.0.0.0 --no-browser --allow-root --port 3000 
-```
-
-The notebook is available at ``http://127.0.0.1:3000/tree``.
-
-## Example usage
-
[…]
