+++ docs/datasets.rst
@@ -1,9 +1,6 @@
 Datasets
 =================
 
-stanscofi.datasets module
--------------------------
-
 .. automodule:: stanscofi.datasets
    :members:
    :undoc-members:
