+++ docs/preprocessing.rst
@@ -1,9 +1,6 @@
 Preprocessing
 =================
 
-stanscofi.preprocessing module
-------------------------------
-
 .. automodule:: stanscofi.preprocessing
    :members:
    :undoc-members:

[…]
