### 2024-01-22 #5973 @Nikoleta-v3
The documentation looks great! Thank you for all your work @clreda.

At this point could you please:

- [ ] If necessary make a new tagged release of your software, and list the version tag of the archived version here. Or let me know if I should just use v2.0.0.
- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
