### 2024-01-22 #5973 @clreda
Dear @Nikoleta-v3, thanks for your praising.

Release: stanscofi [v2.0.0](https://github.com/RECeSS-EU-Project/stanscofi/releases/tag/v2.0.0)
Archive: "A new standard for drug repurposing by collaborative filtering: stanscofi and benchscofi" [v2.0.0](https://zenodo.org/records/10549119) DOI: [10.5281/zenodo.10549119](https://doi.org/10.5281/zenodo.10549119)

Please let me know if anything else is needed.
