+++ paper/paper.bib
@@ -234,18 +234,6 @@
   publisher={John Wiley \& Sons, Ltd Chichester, UK}
 }
 
-@article{luo2018computational,
-  title={Computational drug repositioning using low-rank matrix approximation and randomized algorithms},
-  author={Luo, Huimin and Li, Min and Wang, Shaokai and Liu, Quan and Li, Yaohang and Wang, Jianxin},
-  journal={Bioinformatics},
-  volume={34},
-  number={11},
-  pages={1904--1912},
-  year={2018},
-  doi={10.1093/bioinformatics/bty013},
-  publisher={Oxford University Press}
-}
-
 @article{liang2017lrssl,
   title={LRSSL: predict and interpret drug--disease associations based on data integration using sparse subspace learning},
   author={Liang, Xujun and Zhang, Pengfei and Yan, Lu and Fu, Ying and Peng, Fang and Qu, Lingzhi and Shao, Meiying and Chen, Yongheng and Chen, Zhuchu},
