### 2024-01-22 #5973 @clreda
Dear @Nikoleta-v3 I have just fixed the reference issue mentioned above (commit [d7bdc43](https://github.com/RECeSS-EU-Project/stanscofi/commit/d7bdc43bb66b00f53f7365783dd29de388614358)). The process should end up successfully now.
