### 2024-01-23 #5973 @Kevin-Mattheus-Moerman
@clreda as AEiC I will now help to process this work for acceptance in JOSS. Since this concerns two packages, i.e. `stanscofi` and `benchscofi`, we need to re-do/change a couple of things here first.

Please can you complete the following:
- [ ] Please create a release for the reviewed software for both `stanscofi` and `benchscofi`. Next report back here with the version tags used for these released, for both software packages. I am guessing this will be `v.2.0.0` and `v.2.0.0` ? I will proceed with this just now, but let me know if this needs to be amended. The version tag we'll assign here will be a a collection of the two separated by a comma, e.g. `v.2.0.0, v.2.0.0`.
- [ ] Please create a single merged archive containing both the content for `stanscofi` and `benchscofi`. Next ensure its title matches the paper title, and that you assign the MIT license. For the version tag use the merged one, i.e. `v.2.0.0, v.2.0.0`.

In addition:
- [ ] I'd like to suggest a renaming of the title, it is more common to have the software project name(s) first, e.g.: `stanscofi and benchscofi: A new standard for drug repurposing by collaborative filtering`. This is only a recommendation, but let me know if you are in agreement. If so, please amend the paper and ensure the archive title matches.
