### 2024-01-24 #5973 @clreda
Dear @Kevin-Mattheus-Moerman, thanks for your help.

Release version for ```stanscofi```: ```v2.0.0```
Release version for ```benchscofi```: ```v2.0.0```
Joint archive DOI: Version ```v2.0.0,v2.0.0``` [10.5281/zenodo.10561760](https://doi.org/10.5281/zenodo.10561760)

I agree with your suggestion, and I have changed the title accordingly (commit [c1e0260](https://github.com/RECeSS-EU-Project/stanscofi/commit/c1e02604958f1b652aac0b7725e497e1824a6eb4)) in both the paper and the Zenodo archive.

Please let me know if there is anything else needed.
