### 2024-01-25 d86bac4
<paper> correct definition of sparsity
